web-ww

+++ src/views/containers/shop/guide/GuideListContainer.js
@@ -67,13 +67,18 @@ class GuideListContainer extends Component {
       { title: '店长', dataIndex: 'manager_label', key: 'manager_label'},
       { title: '入职日期', dataIndex: 'entry_time_label', key: 'entry_time_label'},
       { title: '离职日期', dataIndex: 'leave_time_label', key: 'leave_time_label'},
-      { title: '操作', dataIndex: 'id', key: 'id', render:(text, record, index)=>{
-        return (<Button type="primary" shape="circle" icon="delete" size="large" onClick={(e)=>{
-          e.stopPropagation();
-          this.onDelete([record._id])}
-        } />);
-      }}
     ]
+
+    if (this.canOperateDelete()) {
+      columns.push(
+        { title: '操作', dataIndex: 'id', key: 'id', render:(text, record, index)=>{
+          return (<Button type="primary" shape="circle" icon="delete" size="large" onClick={(e)=>{
+            e.stopPropagation();
+            this.onDelete([record._id])}
+          } />);
+        }}
+      );
+    }
 
     return (
       <Root>
@@ -174,6 +179,10 @@ class GuideListContainer extends Component {
   canOperate = () => {
     this.power = commonUtils.getPower(this.props.user, constants.MENU_IDS.shopGuide)
     return this.power && this.power.canOperate;
+  }
+  canOperateDelete = () => {
+    this.power = commonUtils.getPower(this.props.user, constants.MENU_IDS.shopGuide)
+    return this.power && this.power.del;
   }
 }
